requirements.txt
Pillow==11.0.0
uvicorn==0.34.0
fastapi==0.115.8
dotenv==0.9.9
scikit-learn==1.6.1
langchain==0.3.17
requests==2.32.3
regex==2024.11.6
openai==1.60.2
pydantic==2.10.6
pandas==2.2.3
langchain-community==0.3.16
python-multipart==0.0.20
#pdf2image==1.17.0
PyMuPDF==1.25.5